% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{dge_analysis}
\alias{dge_analysis}
\title{This function is for identify DEG from a given matrix}
\usage{
dge_analysis(expr_mat, group)
}
\arguments{
\item{expr_mat}{should a matrix in stead of a data frame, with column for samples and rows for signature(gene/protein/metabolics...)
For transcriptome, a matrix after log transformation is recommended.}

\item{group}{should be a factor whose length is identical to the number of the columns in expr_mat,
describing the group information of each column in expr_mat}
}
\description{
This function is for identify DEG from a given matrix
}
