% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2qcdt.R
\name{exp2qcdt}
\alias{exp2qcdt}
\title{Convert expression table to qc data table.}
\usage{
exp2qcdt(exp_table_file, count_table_file, phenotype_file, result_dir = NULL)
}
\arguments{
\item{exp_table_file}{Expression table file}

\item{phenotype_file}{Phenotype file}

\item{result_dir}{A directory for result files}
}
\description{
Convert expression table to qc data table.
}
