% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_rna_report}
\alias{generate_rna_report}
\title{Generate Quartet RNA report}
\usage{
generate_rna_report(
  qc_result,
  report_template,
  report_dir = NULL,
  report_name = NULL
)
}
\arguments{
\item{qc_result}{list}

\item{report_template}{character}

\item{report_dir}{character}

\item{report_name}{character}
}
\value{
word file
}
\description{
Use calculated RNA result to generate report
}
