% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_RNA_report.R
\name{GenerateRNAReport}
\alias{GenerateRNAReport}
\title{Generate Quartet RNA report}
\usage{
GenerateRNAReport(RNA_result = NULL, doc_file_path = NULL, output_path = NULL)
}
\arguments{
\item{RNA_result}{list}

\item{output_path}{character}

\item{temp_doc_path}{character}
}
\value{
word file
}
\description{
Use calculated RNA result to generate report
}
\examples{
# 加载示例 RNA_result 对象
RNA_result_path <- system.file("extdata", "RNA_result_example.RData", package = "exp2qcdt")
load(RNA_result_path)

# 指定包内文档的路径
doc_file_path_example <- system.file("extdata", "Quartet_temp.docx", package = "exp2qcdt")

# 假设输出路径是当前工作目录
output_path_example <- getwd()

# 运行函数
GenerateRNAReport(RNA_result=RNA_result, doc_file_path=doc_file_path_example, output_path=output_path_example)


}
