% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2qcdt.R
\name{make_directories}
\alias{make_directories}
\title{Make a result directory for exp2qcdt}
\usage{
make_directories(workdir)
}
\arguments{
\item{workdir}{Working directory}
}
\description{
Make a result directory for exp2qcdt
}
