% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_group_output.R
\name{make_performance_plot}
\alias{make_performance_plot}
\title{Make performance figure}
\usage{
make_performance_plot(
  dt_fpkm,
  dt_fpkm_log,
  dt_counts,
  dt_meta,
  result_dir,
  abs_cor_median,
  pt_abs_median_cor,
  ref_data
)
}
\description{
Make performance figure
}
