% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{make_score_figure}
\alias{make_score_figure}
\title{Make the performance score figure}
\usage{
make_score_figure(result_dir, dt_hq_score_scale)
}
\arguments{
\item{result_dir}{A directory for result files}

\item{sample_num}{The number of samples}
}
\description{
Make the performance score figure
}
