% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_rna_report}
\alias{generate_rna_report}
\title{Generate Quartet RNA report}
\usage{
generate_rna_report(
  qc_result,
  report_template,
  report_dir = NULL,
  report_name = NULL
)
}
\arguments{
\item{qc_result}{list}

\item{report_template}{character}

\item{report_dir}{character}

\item{report_name}{character}
}
\description{
Use calculated RNA result to generate report
}
\examples{
# 加载示例 RNA_qc_result 对象
rna_qc_result <- system.file("extdata", "rna_qc_result_example.RData", package = "exp2qcdt")
load(rna_qc_result)

# 指定包内文档的路径
report_template <- system.file("extdata", "quartet_template.docx", package = "exp2qcdt")

# 运行函数
generate_rna_report(qc_result = RNA_result, report_template = report_template)

}
